lifting a bucket to hono
